[#60] Remove extend_chat
