+++ backend/app/api/api_v1/extended/extend_api.py
@@ -1,7 +1,6 @@
 from fastapi import APIRouter
 
 from app.api.api_v1.extended import (
-    extend_chat,
     extend_knowledge_base,
     prompt,
     system_settings,
@@ -14,7 +13,6 @@ api_router.include_router(
     prefix="/knowledge-base",
     tags=["knowledge-base"],
 )
-api_router.include_router(extend_chat.router, prefix="/chat", tags=["chat"])
 api_router.include_router(prompt.router, prefix="/prompt", tags=["prompt"])
 api_router.include_router(
     system_settings.router,
